requirements.txt
Flask
Flask-SQLAlchemy
Flask-SocketIO
flask-socketio